created interface for ExportService
